Anim: change how binding names & assignments work
This commit changes two things:

- How binding names work (addition of a 'display name' RNA property, and
  a replacement of the '??' prefix for unused bindings to 'XX'), and
- Simplification of the C++ API for Bindings, reducing the different
  states the animation system can be in.

Just like ID.name, the Binding.name starts with a two-letter ID code.
This is exposed in RNA as `binding.name`. To access the 'display name',
i.e. the name without the two-letter ID code, use the `.name_display`
RNA property.

Basically there is no more way to "connect" a Binding to an ID without
assigning the Animation. What is possible, however, is to create a
Binding that's already named after the ID, and set up for that ID type.

This reduced the number of ways in which Animation data-blocks can be
assigned to animated IDs, nicely simplifying some things.