Update 2D animation startup file
This makes sure to turn off the unified paint settings for the
brush size in the default `Scene`.

When loading the template, the draw tool uses the size
defined by the brush.
This is needed for GPv3, because we make use of the unified
paint settings, but don't want to use them when loading
the template.