Refactor: Improve blendfile partial write
Instead of temporarily moving datablocks to a new main database, make a
copy so that the original datablocks are never modified.

This is less of a concern for current code, but with brush assets we
want to make more modifications to the datablocks before writing and
then it becomes difficult to restore them to the original state.

The downside is that this is that for heavy scenes, this will make copy
objects and bpy library write significantly slower and memory hungry. So
it's not entirely obvious to me that we should use this always, or only
add the mechanism to be used by brush assets.

Note that this assumes the main argument for BKE_id_copy_ex is the
destination main database, not the source. It's difficult to verify all
data copy functions respect this, but I couldn't find counterexamples,
and for depsgraph COW this is null so there should be no reliance on the
source main database for copying.