Compositor: Make GPU compositor an official feature
Effectively, make GPU compositor available without need to enable
an experimental feature set.

This change does not include changes to the UI, so in order to
switch compositor to GPU mode the N-panel of Compositor editor
is to be used.

It feels like such setting should be either in Performance panel
in render settings, or in a sub-panel in the Post-Processing panel
in output settings. This needs a separate look to find the best
solution.