Fix #121467: EEVEE-Next: Support for Windows/Intel GPU drivers
Intel drivers on Windows platform require all varying interface data
to be written to in the previous stage.

It was detected that the world shader didn't do that as the interface
contains data used for regular objects.

There are several ways how this could be solved:
- Change nodes to use a different path for world materials.
- Include of the draw_model_lib/common_view_lib
- set the failing interface directly.

I chose to set the failing interface directly but I am fine to include
any missing library.