Refactor: Prepare for separate main databases for brush assets
Brush, texture, node and image operators now no longer assume that these
datablocks are part of the global main. Rather they may be exists in a
separate main associated with a brush asset.

While this could be extended to more datablock types, there is currently
no practical benefit. So it doesn't seem worth it unless we support
editing these in a similar way to brush assets, for which there never
may be a reason.

RNA for all data types was modified to remove various assumptions about
global main, and check that when linking datablocks the main databases
are the same. This also makes things safer depsgraph created datablocks
without a main database.