Anim: add failure propagation to more lower-level keying functions
The following keyframing functions now return `SingleKeyingResult`,
which is in turn used for better failure reporting in the higher-level
functions that call them:

- `KeyframeStrip::keyframe_insert()`
- `insert_vert_fcurve()`

As a side effect, this also means that
`rna_KeyframeAnimationStrip_key_insert()` can no longer return an
`FCurve *`, and now instead returns a bool indicating success.

This is part of an ongoing progressive refactor to improve error
messages and failure handling in the keyingframing code.