Refactor: centralize responsibility for "Only Insert Needed"
This makes `insert_vert_fcurve()` responsible for appropriately
handling the `Only Insert Needed` flag.

This is in preparation for the new layered action keyframing code (to
come in a later PR), since we want a common code path for old- and
new-style actions that always handles this flag correctly.  And just
generally this seems like a more sensible place for itto be handled
anyway.