Brush Assets: Implement grease pencil default eraser brush
Instead of a boolean on brushes, this is now a secondary brush pointer
and asset reference. It can be chosen from a new Eraser panel in the
brush tool properties.

It was implemented generically, since this functionality may be useful
to extend to other modes in the future.

This current defaults to Eraser Hard, should be changed to Eraser
Soft once the essentials library has been updated to include this.

TODO: Actually selecting the brush is not werking well, since it only
lists brushes that have been loaded at least once. And for that to work
this also needs to be applied on top of #121597. The proper solution
would be to implement a version of the popover for this, but maybe
that's best left for a separate step.