EEVEE-Next: Shadow: Add resolution clamp
This adds a scaling parameter to avoid near camera pixels
to eat too much shadow resolution. This is more intuitive
than the scale shadow setting and allows reducing the
precision without changing the look of distant shadows.

For local light, this is a slider that clamps the
minimum LOD.

For sun lights, the property sets the minimum pixel
size the shadow can contains. This allows to
remove a lot tilemap close up.

This add versionning code to handle EEVEE-Legacy files
that had fixed resolution per light type.