Vulkan: Fix push constants tests
Push constants test were failing due to using incorrect parameter.
It used the pipeline stage as if it was a shader stage.

When using push constants fallback the uniform was loaded to the
descriptor set after the descriptor set was checked.

Suppress updating push constants of non render graph, when render
graph is active.