Extensions: Install From Disk operator handling both legacy and new
When a user has download an add-on as a zip file, it's not clear in
advance if this is a legacy add-on or a new extensions. So they would
have to use trial and error, or inspect the zip file contents.

This uses a simple heuristic to check if the file is a legacy add-on,
and if so automatically calls that legacy operator instead.