Fix: Studiolight: Use correct defaults for selected studio light
When selecting Scene World with Cycles, the world background will be
missing-texture pink for one frame before showing the correct light.

The issue here appears to be that Cycles queries
rna_View3DShading_selected_studio_light_get, which returns the placeholder
"Default" value which doesn't have a file path, so the world ends up pink.

Immediately afterwards, the UI code calls rna_View3DShading_studio_light_get,
which initializes shading->lookdev_light to the proper default.

The easiest workaround for this is to provide the proper flags when getting
the selected studio light, which will not match the placeholder entry and
therefore return the proper default.