Fix: GPv3: `Layer::get_frame_duration_at`
The `Layer::get_frame_duration_at` was not working for frames
with a fixed duration. While this is not an issue at the
moment (because fixed duration frames are not exposed
yet), this would have been broken in the future.

This fixes the issues, cleans up the code a bit, and also
adds regression tests.