Fix #122150: UI: layout.template_search() crash without an active item
`template_search` was introduced in c20c203b8226

Seems it assumed an active item is always there.
Without an active item, we are running into a crash trying to get the
textbut width.

So solution is to not draw that text button at all until an active item
is chosen.

This also goes in line with what `template_ID` is doing (it also skips
the text button without an active item).