Python API: Expose list of running modal operators
It is accessible via window.modal_operators, and is a collection of
elements of type Operator.

It allows to implement key and operator capture overlay without need
to do low-level ctype style of access to an internal data.

Example code which logs currently running modal operators:
```
import bpy

def log_modal_operators():
    for op in bpy.context.window.modal_operators:
        print(op.name)
    return 0.1

bpy.app.timers.register(log_modal_operators)
```

The motivation is to provide an official API for implementing the
input overlay add-on which is very great to have for tutorials, but
which is not strictly eligible for inclusions to the Extensions
platforms due to ctypes usage. Having an official way of implementing
it for 4.2 LTS, together with the Extensions platform launch sounds
like a very good thing to do.