Fix Image Save operator locking when saving render result
The issue has been introduced by #122105

The image save operator does exactly the thing which was not supported by
the change and was not caught during development or view: it acquires and
releases image buffers while a render result of the Image data-block is
held acquired.

Ideal solution would be to user-count RenderResult, and never hold mutex
for long period of time, but this would be quite a change. Until then the
solution is to relax mutex requirement a bit, allowing image buffers to be
acquired and released. The downside of this approach is that it makes it
possible to acquire a new frame of multi-layer EXR sequence which might
free a render result which has been acquired. This is not a newly introduced
limitation, and it should not affect stability of the previous fixes in
the area.