EEVEE-Next: Improve shadow map filter
This make the filter have a constant width in shadow
space making sure to always filter the correct amount
of pixels.

This uses the slope to bias the disk samples instead of
setting up the disk in tangent space. But we limit the
slope bias to 1 pixel around the receiver and use
normal offset if the angle between light and receiver
is greater than 45 degrees.