Geometry Nodes: rename Curve Spiral to Spiral in Add Node menu
Nodes should have the same name in the Add Node menu as their actual
name, but the Spiral node is called "Curve Spiral". Since there is no
ambiguity with another primitive type called Spiral, it can be renamed
to that.

This also fixes an issue with translation, where the node could not be
translated since its name was never extracted to the translation
files.